Build error messages for dynamic_list questions
Errors for dynamic_list questions need to be linked to the field that is missing: either a primary
question or a follow-up question (depending on the
value of the primary question).
To achieve this, we're putting one extra key into
the ERRORS something dict indicating which field
failed the validation.